+++ Engine/Goal.h
@@ -5,24 +5,24 @@ class Goal
 {
 
 public:
-	Goal(Graphics& gfx, int x, int y);
-	void Draw(Graphics & gfx, int x, int y);
+	Goal(Graphics& gfx, float x, float y);
+	void Draw(Graphics & gfx, float x, float y);
 	bool IsColliding(Player& player);
 	void DrawTest(Graphics& gfx);
-	void Update(int x_in, int y_in);
+	void Update(float x_in, float y_in);
 	//void UpdateColor();
 
 private:
 
 
-	int x;
-	int y;
+	float x;
+	float y;
 	bool pickedUp = false;
 	int timesCaught = 0;
 	Color c{ 255, 255, 255 };
 	bool coloInreaseing = false;
 
 public:
-	static constexpr int width = 20;
-	static constexpr int height = 20;
+	static constexpr float width = 20;
+	static constexpr float height = 20;
 };
